src/requirements.txt
asgiref==3.5.2
Django==4.1.2
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.4
gunicorn==20.1.0
psycopg2-binary==2.9.1

